+++ Source/Core/Cartesian3.js
@@ -69,6 +69,49 @@ define([
     };
 
     /**
+     * Creates a Cartesian3 from three consecutive elements in an array.
+     * @memberof Cartesian3
+     *
+     * @param {Array} values The array whose three consecutive elements correspond to the x, y, and z components, respectively.
+     * @param {Number} [offset=0] The offset into the array of the first element, which corresponds to the x component.
+     * @param {Cartesian3} [result] The object onto which to store the result.
+     *
+     * @return {Cartesian3} The modified result parameter or a new Cartesian3 instance if one was not provided.
+     *
+     * @exception {DeveloperError} values is required.
+     * @exception {DeveloperError} offset + 3 is greater than the length of the array.
+     *
+     * @example
+     * // Create a Cartesian3 with (1.0, 2.0, 3.0)
+     * var v = [1.0, 2.0, 3.0];
+     * var p = Cartesian3.fromArray(v);
+     *
+     * // Create a Cartesian3 with (1.0, 2.0, 3.0) using an offset into an array
+     * var v2 = [0.0, 0.0, 1.0, 2.0, 3.0];
+     * var p2 = Cartesian3.fromArray(v2, 2);
+     */
+    Cartesian3.fromArray = function(values, offset, result) {
+        if (typeof values === 'undefined') {
+            throw new DeveloperError('values is required.');
+        }
+
+        if (offset + 3 > values.length) {
+            throw new DeveloperError('offset + 3 is greater than the length of the array.');
+        }
+
+        offset = defaultValue(offset, 0);
+
+        if (typeof result === 'undefined') {
+            result = new Cartesian3();
+        }
+
+        result.x = values[offset + 0];
+        result.y = values[offset + 1];
+        result.z = values[offset + 2];
+        return result;
+    };
+
+    /**
      * Duplicates a Cartesian3 instance.
      * @memberof Cartesian3
      *
@@ -165,6 +208,32 @@ define([
      */
     Cartesian3.magnitude = function(cartesian) {
         return Math.sqrt(Cartesian3.magnitudeSquared(cartesian));
+    };
+
+    var distanceScratch = new Cartesian3();
+
+    /**
+     * Computes the distance between two points
+     * @memberof Cartesian3
+     *
+     * @param {Cartesian3} left The first point to compute the distance from.
+     * @param {Cartesian3} right The second point to compute the distance to.
+     *
+     * @return {Number} The distance between two points.
+     *
+     * @exception {DeveloperError} left and right are required.
+     *
+     * @example
+     * // Returns 1.0
+     * var d = Cartesian3.distance(new Cartesian3(1.0, 0.0, 0.0), new Cartesian3(2.0, 0.0, 0.0));
+     */
+    Cartesian3.distance = function(left, right) {
+        if ((typeof left === 'undefined') || (typeof right === 'undefined')) {
+            throw new DeveloperError('left and right are required.');
+        }
+
+        Cartesian3.subtract(left, right, distanceScratch);
+        return Cartesian3.magnitude(distanceScratch);
     };
 
     /**